+++ R/get-set-convenience.R
@@ -222,13 +222,8 @@ outer_row_col_value <- function(ht, row, col, value) {
     value <- row
     row <- seq_len(nrow(ht))
     col <- seq_len(ncol(ht))
-  } else if (missing(value)) {
-    value <- col
-    if (!is.matrix(row)) stop("No columns specified, but `row` argument did not evaluate to a matrix")
-    # row is a 2-matrix of row, col vectors;
-    col <- seq(min(row[, 2]), max(row[, 2]))
-    row <- seq(min(row[, 1]), max(row[, 1]))
   }
+
   row <- get_rc_spec(ht, row, 1)
   col <- get_rc_spec(ht, col, 2)
   if (is.logical(row)) row <- which(row)
@@ -286,24 +281,18 @@ set_cell_properties <- function (ht, row, col, ...) {
 #' ```
 #' @return The `ht` object.
 #'
-#' @seealso merge_repeated_rows
+#' @family cell merging
 #'
 #' @export
 #' @examples
+#'
 #' ht <- hux(a = 1:3, b = 1:3)
 #' ht <- set_all_borders(ht, 1)
-#' merge_cells(ht, 1:2, 1:2)
+#' merge_cells(ht, 2:3, 1:2)
+#'
 merge_cells <- function (ht, row, col) {
   assert_that(is_huxtable(ht))
 
-  if (missing(col)) {
-    .Deprecated("Using merge_cells without a `col` argument is deprecated.", package = "huxtable")
-    if (! is.matrix(row)) stop(
-      "No columns specified, but `row` argument did not evaluate to a matrix")
-    # 2-matrix of row, col vectors
-    col <- seq(min(row[, 2]), max(row[, 2]))
-    row <- seq(min(row[, 1]), max(row[, 1]))
-  }
   row <- get_rc_spec(ht, row, 1)
   col <- get_rc_spec(ht, col, 2)
   if (is.logical(row)) row <- which(row)
@@ -315,6 +304,56 @@ merge_cells <- function (ht, row, col) {
   rs <- diff(range(row)) + 1
   colspan(ht)[mr, mc] <- cs
   rowspan(ht)[mr, mc] <- rs
+
+  ht
+}
+
+
+#' Merge cells across rows or down columns
+#'
+#' `merge_across` creates multicolumn cells within each row. `merge_down` creates
+#' multirow cells within each column.
+#'
+#' @inherit left_border params
+#'
+#' @return The `ht` object.
+#'
+#' @export
+#'
+#' @family cell merging
+#'
+#' @examples
+#'
+#' ht <- as_hux(matrix(1:12, 4, 3, byrow = TRUE))
+#' ht <- set_all_borders(ht, 1)
+#' merge_across(ht, 2:4, 2:3)
+#' merge_down(ht, 2:4, 2:3)
+#'
+merge_across <- function (ht, row, col) {
+  assert_that(is_huxtable(ht))
+
+  row <- get_rc_spec(ht, row, 1)
+  col <- get_rc_spec(ht, col, 2)
+  if (is.logical(row)) row <- which(row)
+  if (is.logical(col)) col <- which(col)
+
+  for (r in row) ht <- merge_cells(ht, r, col)
+
+  ht
+}
+
+
+#' @rdname merge_across
+#' @export
+merge_down <- function (ht, row, col) {
+  assert_that(is_huxtable(ht))
+
+  row <- get_rc_spec(ht, row, 1)
+  col <- get_rc_spec(ht, col, 2)
+  if (is.logical(row)) row <- which(row)
+  if (is.logical(col)) col <- which(col)
+
+  for (cl in col) ht <- merge_cells(ht, row, cl)
 
   ht
 }